Split out FDefaultCommunityTab
Also move the option to the General > General section.
