+++ src/html/options.html
@@ -73,6 +73,7 @@
                     <div data-dynamic="send_age_info"></div>
                     <div data-dynamic="openinnewtab"></div>
                     <div data-dynamic-select="installsteam"></div>
+                    <div data-dynamic-select="community_default_tab"></div>
                 </div>
 
                 <div class="section js-section">
@@ -322,7 +323,6 @@
                 <div class="section js-section">
                     <h2 data-locale-text="options.general" class="js-section-head">General</h2>
                     <div data-dynamic-select="show_profile_link_images"></div>
-                    <div data-dynamic-select="community_default_tab"></div>
                     <div data-dynamic="showallachievements"></div>
                     <div data-dynamic="showallstats"></div>
                     <div data-dynamic="hidespamcomments" class="parent_option"></div>
