+++ src/js/Content/Features/Community/CCommunityBase.js
@@ -3,6 +3,7 @@ import FHideTrademarks from "../Common/FHideTrademarks";
 import FBackToTop from "../Common/FBackToTop";
 import FDisableLinkFilter from "../Common/FDisableLinkFilter";
 import FKeepSSACheckboxState from "../Common/FKeepSSACheckboxState";
+import FDefaultCommunityTab from "../Common/FDefaultCommunityTab";
 import FConfirmDeleteComment from "./FConfirmDeleteComment";
 import FHideSpamComments from "./FHideSpamComments";
 import FFavoriteEmoticons from "./FFavoriteEmoticons";
@@ -16,6 +17,7 @@ export class CCommunityBase extends Context {
             FBackToTop,
             FDisableLinkFilter,
             FKeepSSACheckboxState,
+            FDefaultCommunityTab,
             FConfirmDeleteComment,
             FHideSpamComments,
             FFavoriteEmoticons,
